+++ core/interfaces.go
@@ -20,7 +20,7 @@ type Services interface {
 
 	SerGetAppConfigs(version string) ([]model.ApplicationConfigs, error)
 	SerGetAppConfig(id string) (*model.ApplicationConfigs, error)
-	SerCreateAppConfig(appConfig model.ApplicationConfigs) (*model.ApplicationConfigs, error)
+	SerCreateAppConfig(version string, appID string, data map[string]interface{}) (*model.ApplicationConfigs, error)
 	SerUpdateAppConfig(id string, version string, data map[string]interface{}) error
 	SerDeleteAppConfig(id string) error
 }
@@ -43,11 +43,12 @@ type Administration interface {
 	AdmGetApplication(ID string) (*model.Application, error)
 	AdmGetApplications() ([]model.Application, error)
 
-	AdmCreateApplicationPermission(name string, appID string) (*model.ApplicationPermission, error)
+	AdmCreatePermission(name string, serviceIDs []string) (*model.Permission, error)
+	AdmUpdatePermission(name string, serviceIDs *[]string) (*model.Permission, error)
 
 	AdmCreateApplicationRole(name string, appID string, description string, permissionNames []string) (*model.ApplicationRole, error)
 
-	AdmGrantAccountPermissions(accountID string, appID string, permissionNames []string) error
+	AdmGrantAccountPermissions(accountID string, permissionNames []string) error
 	AdmGrantAccountRoles(accountID string, appID string, roleIDs []string) error
 }
 
@@ -70,17 +71,17 @@ type Storage interface {
 	DeleteAccount(id string) error
 	UpdateAccountPreferences(accountID string, preferences map[string]interface{}) error
 	UpdateProfile(accountID string, profile *model.Profile) error
-	InsertAccountPermissions(accountID string, appID string, permissions []model.ApplicationPermission) error
+	InsertAccountPermissions(accountID string, permissions []model.Permission) error
 	InsertAccountRoles(accountID string, appID string, roles []model.ApplicationRole) error
 
 	CreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 	GetGlobalConfig() (*model.GlobalConfig, error)
 	SaveGlobalConfig(setting *model.GlobalConfig) error
 
-	FindApplicationPermissionsByName(names []string, appID string) ([]model.ApplicationPermission, error)
-	InsertApplicationPermission(item model.ApplicationPermission) error
-	UpdateApplicationPermission(item model.ApplicationPermission) error
-	DeleteApplicationPermission(id string) error
+	FindPermissionsByName(names []string) ([]model.Permission, error)
+	InsertPermission(item model.Permission) error
+	UpdatePermission(item model.Permission) error
+	DeletePermission(id string) error
 
 	FindApplicationRoles(ids []string, appID string) ([]model.ApplicationRole, error)
 	InsertApplicationRole(item model.ApplicationRole) error