+++ driver/web/adapter.go
@@ -71,6 +71,7 @@ func (we Adapter) Start() {
 	servicesSubRouter.HandleFunc("/auth/login-url", we.wrapFunc(we.servicesApisHandler.authLoginURL, nil)).Methods("POST") //Requires API key in request
 	servicesSubRouter.HandleFunc("/auth/refresh", we.wrapFunc(we.servicesApisHandler.authRefresh, nil)).Methods("POST")    //Requires API key in request
 	servicesSubRouter.HandleFunc("/auth/verify", we.wrapFunc(we.servicesApisHandler.verifyCode, nil)).Methods("GET")       //Public (validates code)
+	servicesSubRouter.HandleFunc("/auth/account-exists", we.wrapFunc(we.servicesApisHandler.accountExists, nil)).Methods("POST")
 	servicesSubRouter.HandleFunc("/auth/authorize-service", we.wrapFunc(we.servicesApisHandler.authAuthorizeService, we.auth.servicesUserAuth)).Methods("POST")
 	servicesSubRouter.HandleFunc("/auth/service-regs", we.wrapFunc(we.servicesApisHandler.getServiceRegistrations, we.auth.servicesAuth)).Methods("GET")
 	servicesSubRouter.HandleFunc("/account", we.wrapFunc(we.servicesApisHandler.deleteAccount, we.auth.servicesUserAuth)).Methods("DELETE")
@@ -81,17 +82,21 @@ func (we Adapter) Start() {
 	servicesSubRouter.HandleFunc("/test", we.wrapFunc(we.servicesApisHandler.getTest, nil)).Methods("GET") //Public
 
 	// appconfigs
-	servicesSubRouter.HandleFunc("/app/configs)", we.wrapFunc(we.servicesApisHandler.getAppConfigs, nil)).Methods("GET")
-	servicesSubRouter.HandleFunc("/app/configs)", we.wrapFunc(we.servicesApisHandler.createAppConfig, nil)).Methods("POST")
-	servicesSubRouter.HandleFunc("/app/configs/{id})", we.wrapFunc(we.servicesApisHandler.getAppConfig, nil)).Methods("GET")
-	servicesSubRouter.HandleFunc("/app/configs/{id})", we.wrapFunc(we.servicesApisHandler.updateAppConfig, nil)).Methods("PUT")
-	servicesSubRouter.HandleFunc("/app/configs/{id})", we.wrapFunc(we.servicesApisHandler.deleteAppConfig, nil)).Methods("DELETE")
+	servicesSubRouter.HandleFunc("/app/configs", we.wrapFunc(we.servicesApisHandler.getAppConfigs, nil)).Methods("GET")
+	servicesSubRouter.HandleFunc("/app/configs", we.wrapFunc(we.servicesApisHandler.createAppConfig, nil)).Methods("POST")
+	servicesSubRouter.HandleFunc("/app/configs/{id}", we.wrapFunc(we.servicesApisHandler.getAppConfig, nil)).Methods("GET")
+	servicesSubRouter.HandleFunc("/app/configs/{id}", we.wrapFunc(we.servicesApisHandler.updateAppConfig, nil)).Methods("PUT")
+	servicesSubRouter.HandleFunc("/app/configs/{id}", we.wrapFunc(we.servicesApisHandler.deleteAppConfig, nil)).Methods("DELETE")
 	///
 
 	///admin ///
 	adminSubrouter := subRouter.PathPrefix("/admin").Subrouter()
 	adminSubrouter.HandleFunc("/test", we.wrapFunc(we.adminApisHandler.getTest, we.auth.adminAuth)).Methods("GET")
 	adminSubrouter.HandleFunc("/test-model", we.wrapFunc(we.adminApisHandler.getTestModel, we.auth.adminAuth)).Methods("GET")
+
+	adminSubrouter.HandleFunc("/auth/login", we.wrapFunc(we.adminApisHandler.adminLogin, nil)).Methods("POST")
+	adminSubrouter.HandleFunc("/auth/login-url", we.wrapFunc(we.adminApisHandler.adminLoginURL, nil)).Methods("POST")
+	adminSubrouter.HandleFunc("/auth/refresh", we.wrapFunc(we.adminApisHandler.adminRefresh, nil)).Methods("POST")
 
 	adminSubrouter.HandleFunc("/global-config", we.wrapFunc(we.adminApisHandler.createGlobalConfig, we.auth.adminAuth)).Methods("POST")
 	adminSubrouter.HandleFunc("/global-config", we.wrapFunc(we.adminApisHandler.getGlobalConfig, we.auth.adminAuth)).Methods("GET")
@@ -106,7 +111,9 @@ func (we Adapter) Start() {
 	adminSubrouter.HandleFunc("/applications/{id}", we.wrapFunc(we.adminApisHandler.getApplication, we.auth.adminAuth)).Methods("GET")
 	adminSubrouter.HandleFunc("/applications", we.wrapFunc(we.adminApisHandler.getApplications, we.auth.adminAuth)).Methods("GET")
 
-	adminSubrouter.HandleFunc("/application-permissions", we.wrapFunc(we.adminApisHandler.createApplicationPermission, we.auth.adminAuth)).Methods("POST")
+	adminSubrouter.HandleFunc("/permissions", we.wrapFunc(we.adminApisHandler.createPermission, we.auth.adminAuth)).Methods("POST")
+	adminSubrouter.HandleFunc("/permissions", we.wrapFunc(we.adminApisHandler.updatePermission, we.auth.adminAuth)).Methods("PUT")
+
 	adminSubrouter.HandleFunc("/application-roles", we.wrapFunc(we.adminApisHandler.createApplicationRole, we.auth.adminAuth)).Methods("POST")
 	adminSubrouter.HandleFunc("/application-api-keys", we.wrapFunc(we.adminApisHandler.getApplicationAPIKeys, we.auth.adminAuth)).Methods("GET")