+++ CHANGELOG.md
@@ -6,14 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Merge client and Profile BB profiles and preferences [#228](https://github.com/rokwire/core-building-block/issues/228)
+
+## [1.9.0] - 2021-11-01
+### Added
+- Disable email verification [#280](https://github.com/rokwire/core-building-block/issues/280)
+
+### Changed
+- Dissociate permissions from applications [#207](https://github.com/rokwire/core-building-block/issues/207)
+
+## [1.8.0] - 2021-10-27
+### Added
+- Prepare Core for Admin app integration [#247](https://github.com/rokwire/core-building-block/issues/247)
+- Expose does account exist services API [#255](https://github.com/rokwire/core-building-block/issues/255)
+
+### Fixed
+- MongoDB ChangeStream Watch() does not recover [#259](https://github.com/rokwire/core-building-block/issues/259)
 
 ## [1.7.0] - 2021-10-25
 ### Added
 - Handle API key validation for non-anonymous users [#244](https://github.com/rokwire/core-building-block/issues/244)
 - Implement logins sessions [#172](https://github.com/rokwire/core-building-block/issues/172) 
-
-### Fixed
-- MongoDB ChangeStream Watch() does not recover [#259](https://github.com/rokwire/core-building-block/issues/259)
 
 ## [1.6.0] - 2021-10-19
 ### Added
@@ -21,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.5.0] - 2021-10-15
 ### Fixed
-- Permission authorization failing on all endpoints in Docker[#239](https://github.com/rokwire/core-building-block/issues/239)
+- Permission authorization failing on all endpoints in Docker [#239](https://github.com/rokwire/core-building-block/issues/239)
 
 ### Changed
 - Switch to ROKWIRE open source libraries [#232](https://github.com/rokwire/core-building-block/issues/232)
@@ -114,7 +128,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.7.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/rokwire/core-building-block/compare/v1.8.0...v1.9.0
+[1.8.0]: https://github.com/rokwire/core-building-block/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/rokwire/core-building-block/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/rokwire/core-building-block/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/rokwire/core-building-block/compare/v1.4.0...v1.5.0